bitch

+++ ft_handledigits.c
@@ -18,7 +18,7 @@ void	ft_handledigits(f_list *params, void *arg, int val)
 	if (params->conversion == 'x' || params->conversion == 'X')
 		ft_printhex(params->conversion, arg, val);
 	if (params->conversion == 'o' || params->conversion == 'O')
-		ft_printcocto((int)arg, val);
+		ft_printcocto(arg, val);
 	if (params->conversion == 'b')
 		ft_printbinary((int)arg, val);
 //	if (params->conversion == 'k')
